requirements.txt
pypng
tensorflow >= 1.0.1
PyWavelets
scipy
matplotlib
scikit-learn
cvxopt